# openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Invoice Manager API docs",
        "version": "1.0.0"
    },
    "components": {
        "securitySchemas": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
        "schemas": {
            "Customer": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "number": {
                        "type": "number"
                    },
                    "name": {
                        "type": "string"
                    },
                    "streetAddress": {
                        "type": "string"
                    },
                    "postalCode": {
                        "type": "string"
                    },
                    "city": {
                        "type": "string"
                    },
                    "country": {
                        "type": "string"
                    },
                    "identificationNumber": {
                        "type": "string"
                    },
                    "taxExempt": {
                        "type": "boolean"
                    }
                }
            },
            "Invoice": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "customerId": {
                        "type": "string"
                    },
                    "number": {
                        "type": "number"
                    },
                    "paid": {
                        "type": "boolean"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "dueAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "currency": {
                        "type": "string"
                    }
                }
            },
            "InvoiceLine": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "invoiceId": {
                        "type": "string"
                    },
                    "lineText": {
                        "type": "string"
                    },
                    "amount": {
                        "type": "number"
                    },
                    "quantity": {
                        "type": "number"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "InvoicePayment": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "invoiceId": {
                        "type": "string"
                    },
                    "paymentMethod": {
                        "type": "string"
                    },
                    "amount": {
                        "type": "number"
                    },
                    "externalPaymentId": {
                        "type": "number"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "Settings": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "streetAddress": {
                        "type": "string"
                    },
                    "postalCode": {
                        "type": "string"
                    },
                    "city": {
                        "type": "string"
                    },
                    "country": {
                        "type": "string"
                    },
                    "identificationNumber": {
                        "type": "string"
                    },
                    "vatPayer": {
                        "type": "boolean"
                    },
                    "preferredCurrency": {
                        "type": "string"
                    },
                    "additionalText": {
                        "type": "string"
                    }
                }
            }
        }
    },
    "paths": {
        "/api/v1/customers": {
            "get": {
                "description": "Lists all customers (paginated)",
                "tags": [
                    "Customer"
                ],
                "responses": {
                    "200": {
                        "description": "Returns customers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Customer"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a customer",
                "tags": [
                    "Customer"
                ],
                "requestBody": {
                    "description": "Customer created",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Customer"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/customers/{id}": {
            "delete": {
                "description": "Delete a customer",
                "tags": [
                    "Customer"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Customer ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted successfully"
                    }
                }
            }
        },
        "/api/v1/invoices": {
            "get": {
                "description": "Lists all invoices(paginated)",
                "tags": [
                    "Invoice"
                ],
                "responses": {
                    "200": {
                        "description": "Returns invoices",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Invoice"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create an invoice",
                "tags": [
                    "Invoice"
                ],
                "requestBody": {
                    "description": "Invoice created",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Invoice"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/invoices/{id}": {
            "get": {
                "description": "Get an invoice by ID",
                "tags": [
                    "Invoice"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Invoice ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns an invoice",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Invoice"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update an invoice",
                "tags": [
                    "Invoice"
                ],
                "requestBody": {
                    "description": "Invoice object",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Invoice"
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Invoice ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns the updated invoice",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Invoice"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/invoices/{id}/lines": {
            "post": {
                "description": "Insert an invoice line",
                "tags": [
                    "Invoice"
                ],
                "requestBody": {
                    "description": "Invoice line object",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InvoiceLine"
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Invoice ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns the created invoice line",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InvoiceLine"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/invoices/{id}/payments": {
            "post": {
                "description": "Insert an invoice payment",
                "tags": [
                    "Invoice"
                ],
                "requestBody": {
                    "description": "Invoice payment object",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InvoicePayment"
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Invoice ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns the created invoice payment",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InvoicePayment"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/settings": {
            "get": {
                "description": "Retrieves user settings",
                "tags": [
                    "Settings"
                ],
                "responses": {
                    "200": {
                        "description": "Returns user settings",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Settings"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update user settings",
                "tags": [
                    "Settings"
                ],
                "requestBody": {
                    "description": "User settings",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Settings"
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": []
}